moving method to BaseDataset, now the method takes care of changing the dataset to streaming and then revert to previous state
